SpacePong 2.0 completo. Primeira release funcional, em progresso de aprimoramento

+++ src/utils/Location.java
@@ -2,28 +2,28 @@ package utils;
 
 public class Location {
   
-  private int x, y;
+  private double x, y;
 
   public Location() { }
 
-  public Location(int x, int y) {
+  public Location(double x, double y) {
     this.x = x;
     this.y = y;
   }
-  
-  public int getX() {
+
+  public double getX() {
     return x;
   }
 
-  public void setX(int x) {
+  public void setX(double x) {
     this.x = x;
   }
 
-  public int getY() {
+  public double getY() {
     return y;
   }
 
-  public void setY(int y) {
+  public void setY(double y) {
     this.y = y;
   }
